adds casts to enso config files

+++ config/enso/auth.php
@@ -1,12 +1,12 @@
 <?php
 
 return [
-    'maxLoginAttempts' => env('LOGIN_ATTEMPTS_PER_MINUTE', 5),
+    'maxLoginAttempts' => (int) env('LOGIN_ATTEMPTS_PER_MINUTE', 5),
     'password' => [
-        'lifetime' => env('PASSWORD_LIFETIME', 0),
-        'minLength' => env('PASSWORD_LENGTH', 6),
-        'minUpperCase' => env('PASSWORD_UPPER_CASE_CHARACTERS', 0),
-        'minNumeric' => env('PASSWORD_NUMERIC_CHARACTERS', 0),
-        'minSpecial' => env('PASSWORD_SPECIAL_CHARACTERS', 0),
+        'lifetime' => (int) env('PASSWORD_LIFETIME', 0),
+        'minLength' => (int) env('PASSWORD_LENGTH', 6),
+        'minUpperCase' => (int) env('PASSWORD_UPPER_CASE_CHARACTERS', 0),
+        'minNumeric' => (int) env('PASSWORD_NUMERIC_CHARACTERS', 0),
+        'minSpecial' => (int) env('PASSWORD_SPECIAL_CHARACTERS', 0),
     ],
 ];
